Prelimimar version of the Skew-Symmetric Split form scheme.
	-> Select using the split form DG by setting:
		* inviscid discretization = Split-Form
in the control file.

	-> Morinishi, Ducros, Kennedy-Gruber and Pirozzoli schemes have
been implemented. It is selected in the control file as:
		* split form = Ducros / ...

	-> Split form schemes does only work with Gauss-Lobatto points.
Hence, if Gauss points are selected (recall that they can be selected in
the control file as:
		* discretization nodes = Gauss / Gauss-Lobatto.
they will be switched automatically to Gauss-Lobatto points

	-> The last modification requires controlVariables to be
editable, so a procedure in FTDictionaryClass.f90 to remove existing objects
was implemented.