I believe the 3D geometries bug is fully fixed
-> Finally there was no other way than computing metric terms on
Chebyshev-Lobatto points to ensure continuity accross elements

-> Added the Chebyshev-Gauss-Lobatto points to the NodalStorage class.
It would be nice to get back to all procedures that use CGL points and
use the built-in ones rather than construct local versions.

-> Added -fPIC to some important libraries. It is important to check on
buildbot if has any drastic effect on the efficiency. This was required
for the MeshConsistency test case. If they really are noticeable, it
will be changed to a different binary, similar to the rest of the
components test.