Add SVV-LES with alpha prop. to mu
The SVV can be coupled with a Smagorinsky LES viscosity by specifying
"SVV viscosity = Smagorinsky" in the control file. In addition, the
value of alpha can be set to be proportional to mu by adding the option "SVV
alpha/mu ratio" in the control file ("SVV alpha viscosity" has priority
over this option).