Improve k-means and GMM clustering
Both methods share the same API now. They also store by default
the clustering information of the last fit, using it to improve the
speed in the following call. In the case of the GMM, clusters removed in
the last fit are created again with default values. This also means that
the values of tau are lost, adding some randomness to the algorithm.