First implementation of a common solution storage
The solution is stored in small arrays scattered over all the elements.
Additionally, a global vector also exists to enable implicit time integration.
This commit defines the element-local arrays as pointers to the corresponding
regions in the global vector, avoiding allocations and automatically coupling
local and global representations of the solution.

Still WIP as some features are missing, namely p-adaptation and the creation of
individual elements (not attached to the global storage).